allow administrator to specify sections on the command line
as requested by Bob West

With both cloginrc.php and routerdb.php, they both simply believe what is in the devices table and will overwrite files based on that information.  What I would prefer is to be able to specify on the command-line which groups to operate on.  E.g. in crontab I would would have:

0 6  * * * php /var/www/fetid/cloginrc.php BCB ABC XYZ
10 6 * * * php /var/www/fetid/routerdb.php BCB ABC XYZ

and this would limit the script to writing groups BCB, ABC and XYZ.  This way somebody adding a new group won't clobber any existing groups without the administrator of the machine doing so on purpose.